moves statistics to its own page
