+++ frontend/src/PagesAdmin/RecruitmentStatistics/RecruitmentStatistics/index.ts
